WIP Retira o formulário de cadastro
